third_party/WebKit/LayoutTests/platform/mac/editing/caret/caret-height-expected.txt

Tests whether explicitly set caret style is retained on performing text selection

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL cursorLine1 should be 14. Was 13.
PASS successfullyParsed is true

TEST COMPLETE

 
